+++ libs/zhi-device/CHANGELOG.md
@@ -1,5 +1,11 @@
 # zhi-device
 
+## 1.1.6
+
+### Patch Changes
+
+- add hash query string
+
 ## 1.1.5
 
 ### Patch Changes