+++ libs/zhi-xmlrpc-middleware/CHANGELOG.md
@@ -1,5 +1,15 @@
 # zhi-xmlrpc-middleware
 
+## 0.2.6
+
+### Patch Changes
+
+- add hash query string
+- Updated dependencies
+  - zhi-lib-base@0.2.6
+  - zhi-common@1.9.6
+  - zhi-device@1.1.6
+
 ## 0.2.5
 
 ### Patch Changes