+++ tools/esbuild-config-custom/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @terwer/esbuild-config-custom
 
+## 1.0.12
+
+### Patch Changes
+
+- add hash query string
+
 ## 1.0.11
 
 ### Patch Changes