+++ tools/eslint-config-custom/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @terwer/eslint-config-custom
 
+## 1.3.6
+
+### Patch Changes
+
+- add hash query string
+
 ## 1.3.5
 
 ### Patch Changes